+++ src/main/kotlin/com/example/demo/controllers/UserController.kt
@@ -1,5 +1,6 @@
-package com.example.demo
+package com.example.demo.controllers
 
+import com.example.demo.repositories.UserRepository
 import org.springframework.http.HttpStatus
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
