+++ src/main/kotlin/com/example/demo/entities/User.kt
@@ -1,4 +1,4 @@
-package com.example.demo
+package com.example.demo.entities
 
 import javax.persistence.Entity
 import javax.persistence.GeneratedValue
