+++ src/main/kotlin/com/example/demo/repositories/UserRepository.kt
@@ -1,5 +1,6 @@
-package com.example.demo
+package com.example.demo.repositories
 
+import com.example.demo.entities.User
 import org.springframework.data.repository.CrudRepository
 
 interface UserRepository : CrudRepository<User, Long> {
