+++ src/test/kotlin/com/example/demo/RepositoriesTests.kt
@@ -1,16 +1,18 @@
 package com.example.demo
 
+import com.example.demo.entities.User
+import com.example.demo.repositories.UserRepository
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.autoconfigure.orm.jpa.AutoConfigureTestEntityManager
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest
 import org.springframework.boot.test.autoconfigure.orm.jpa.TestEntityManager
 
 @DataJpaTest
 class RepositoriesTests @Autowired constructor (
     val entityManager: TestEntityManager,
-    val userRepository: UserRepository) {
+    val userRepository: UserRepository
+) {
 
     @Test
     fun `When findByLogin then return User`() {
